requirements.txt
numpy>=1.21.0
Pillow>=9.0.0
PyPDF2>=3.0.0
PyMuPDF>=1.19.0
pyinstaller>=6.5.0
pytest>=8.0.0
pytest-cov>=4.1.0